requirements.txt
astroid==3.3.11
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
dill==0.4.1
exceptiongroup==1.3.1
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.1.0
isort==6.1.0
mccabe==0.7.0
packaging==26.0
platformdirs==4.4.0
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.19.2
pylint==3.3.9
pytest==8.4.2
pytest-base-url==2.1.0
pytest-playwright==0.7.1
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
text-unidecode==1.3
tomli==2.4.0
tomlkit==0.14.0
typing_extensions==4.15.0
urllib3==2.6.3
zipp==3.23.0
